set to original location, changed by mistake

+++ src/main/gethNodeController.js
@@ -71,8 +71,12 @@ GethNodeController.prototype.start = function (event) {
   if (!fs.existsSync(gethPath)) {
     fs.mkdirSync(gethPath)
   }
-  const staticNodesPath = path.join(gethPath, 'static-nodes.json')
-  fs.writeFileSync(staticNodesPath, JSON.stringify(PEER_NODES, null, 4))
+
+  const staticNodesPath = path.join(gethPath, 'geth', 'static-nodes.json')
+  if (!fs.existsSync(staticNodesPath)) {
+    fs.writeFileSync(staticNodesPath, JSON.stringify(PEER_NODES, null, 4))
+  }
+
   this.gethProcess = spawn(this.gethExecutablePath, [
     '--syncmode=light',
     '--cache=512',
